add speculative support for Windows
I don't have a Windows install so this isn't something that I can easily test, so for now speculate based on what other CinderBlock settings are and hopefully James can clean things up. 🙏

Sadly the CinderBlock spec doesn't allow for an architecture attribute so while the Oculus SDK is available for both Win32 and x64, we have to pick one. 💩